+++ CHANGELOG.md
@@ -8,17 +8,39 @@ For each release, the changelog describes what was added, changed, fixed, or rem
 
 At the end of this file, the release history is provided with links to previous releases.
 
-## 1.1.1patch - 2026-08-25
+## 1.2.0 - 2026-08-25
 
-This is `Patch Version`, so only will change something or fix some bugs.
+Data Update
+
+### Runtime
+
+- Added boolean (true, false) support.
+- Added none support.
+- Added is_integer().
+- Added is_float().
+
+### Parser
+
+- Can sort Integer, Float, String, Boolean, and none.
 
 ### Editor
 
-- Added input() to CompletionItems.
-- Hover built-in function tag `(Built-in Function)` became `(Function)`.
+- Added is_integer() to CompletionItems.
+- Added is_float() to CompletionItems.
+- Added none to CompletionItems.
+- Added true to CompletionItems.
+- Added false to CompletionItems.
+- Added hover information for is_integer().
+- Added hover information for is_float().
+- Added highlight for is_integer().
+- Added highlight for is_float().
+- Added highlight for true.
+- Added highlight for false.
+- Added highlight for none.
 
 ## Releases
 
+- [1.2.0](https://github.com/qzou1222-alt/GameScript/blob/v1.2.0/CHANGELOG.md)
 - [1.1.1patch](https://github.com/qzou1222-alt/GameScript/blob/v1.1.1patch/CHANGELOG.md)   `patch`
 - [1.1.0](https://github.com/qzou1222-alt/GameScript/blob/v1.1.0/CHANGELOG.md)
 - [1.0.0](https://github.com/qzou1222-alt/GameScript/blob/v1.0.0/CHANGELOG.md)